+++ types/Carousel/plugins/index.d.ts
@@ -1,2 +1,5 @@
 import { PluginsType } from "../types";
-export declare const Plugins: PluginsType;
+export type * from "./Dots/Dots";
+export type * from "./Navigation/Navigation";
+export type * from "./Sync/Sync";
+export declare const CarouselPlugins: PluginsType;