+++ types/l10n/Carousel/en.d.ts
@@ -1,5 +1,0 @@
-export declare const en: {
-    NEXT: string;
-    PREV: string;
-    GOTO: string;
-};