+++ types/shared/utils/toggleClass.d.ts
@@ -1,1 +1,1 @@
-export declare const toggleClass: (el: HTMLElement, classes?: string, condition?: boolean) => void;
+export declare const toggleClass: (el: HTMLElement | null, classes?: string, condition?: boolean) => void;